move to its own file rather than sitting in prefs
